Scenario runner (#3)
Summary:

This commit adds a scenario runner and demo script which will run an agent through a set number of rounds in a set number of scenarios, using an EnvironmentGenerator to create a new Environment for each scenario. There is a demo script that runs a randomAgent in a CipherPuzzle environment.

Test Plan:

I added a test for the simulator and manually checked the demo script.